Moved "join us" controller into a shared folder
